gestionPedidos/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Markdown==3.7
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
pytest==7.4.0
pytest-django==4.5.2
xhtml2pdf
drf_yasg
whiteNoise
gunicorn
django-cors-headers
